requirements.txt
peewee
python-dotenv
pymysql